Conflicts:
	app/Http/Controllers/Auth/RegisterController.php
	app/Http/Controllers/Users/AccountController.php
	app/Models/User/User.php
	app/Services/UserService.php
	composer.json

+++ app/Services/CurrencyService.php
@@ -55,6 +55,8 @@ class CurrencyService extends Service
 
             $currency = Currency::create($data);
 
+            if(!logAdminAction($user, 'Created Currency', 'Created '.$currency->displayName)) throw new \Exception("Failed to log admin action.");
+
             if ($icon) $this->handleImage($icon, $currency->currencyIconPath, $currency->currencyIconFileName);
             if ($image) $this->handleImage($image, $currency->currencyImagePath, $currency->currencyImageFileName);
 
@@ -99,6 +101,8 @@ class CurrencyService extends Service
             }
 
             $currency->update($data);
+
+            if(!logAdminAction($user, 'Updated Currency', 'Updated '.$currency->displayName)) throw new \Exception("Failed to log admin action.");
 
             if ($icon) $this->handleImage($icon, $currency->currencyIconPath, $currency->currencyIconFileName);
             if ($image) $this->handleImage($image, $currency->currencyImagePath, $currency->currencyImageFileName);
@@ -170,7 +174,7 @@ class CurrencyService extends Service
      * @param  \App\Models\Currency\Currency  $currency
      * @return bool
      */
-    public function deleteCurrency($currency)
+    public function deleteCurrency($currency, $user)
     {
         DB::beginTransaction();
 
@@ -178,7 +182,7 @@ class CurrencyService extends Service
 
             if(DB::table('loots')->where('rewardable_type', 'Currency')->where('rewardable_id', $currency->id)->exists()) throw new \Exception("A loot table currently distributes this currency as a potential reward. Please remove the currency before deleting it.");
             if(DB::table('prompt_rewards')->where('rewardable_type', 'Currency')->where('rewardable_id', $currency->id)->exists()) throw new \Exception("A prompt currently distributes this currency as a reward. Please remove the currency before deleting it.");
-            if(DB::table('shop_stock')->where('currency_id', $currency->id)->exists()) throw new \Exception("A shop currently requires this currency to purchase an item. Please change the currency before deleting it.");
+            if(DB::table('shop_stock')->where('currency_id', $currency->id)->exists()) throw new \Exception("A shop currently requires this currency to purchase an currency. Please change the currency before deleting it.");
             // Disabled for now due to issues with JSON lookup with older mysql versions/mariaDB
             // if(DB::table('items')->where('data->resell', $currency->id)->exists()) throw new \Exception("An item currently uses this currency for its resale value. Please change the resale information before deleting this currency.");
 
@@ -186,6 +190,8 @@ class CurrencyService extends Service
             // The reason this is allowed is that in instances where event currencies
             // are created for temporary use, it would be inconvenient to have to manually
             // remove them from user accounts before deleting the base currency.
+
+            if(!logAdminAction($user, 'Deleted Currency', 'Deleted '.$currency->name)) throw new \Exception("Failed to log admin action.");
 
             UserCurrency::where('currency_id', $currency->id)->delete();
             CharacterCurrency::where('currency_id', $currency->id)->delete();
